Circular Queue Array
In this technique front and queue initialized 0 not -1. Where ever front point at location that location must be empty or don't bother that location and must that space empty. So the array size 6 then assume that we had space for storing only 5 elements. Array is not a circular front and rear both will acting circularly.

For moving circular concept used in (mod %) and assume size is 4

rear     =      (rear + 1) % size
0         =      (0 + 1) % 4 = 1
1         =      (1 + 1) % 4 = 2
2         =      (2 + 1) % 4 = 3
3         =      (3 + 1) % 4 = 0
0         =      (0 + 1) % 4 =  1 .... and again we go further it will act like circularly

front     =      (front     + 1) % size
0         =      (0 + 1) % 4 = 1
1         =      (1 + 1) % 4 = 2
2         =      (2 + 1) % 4 = 3
3         =      (3 + 1) % 4 = 0
0         =      (0 + 1) % 4 =  1 .... and again we go further it will act like circularly