Priority Queue using Three Queue
In this program, one queue structure having priority and other queue is simple. Concept of this priority queue is insert data with their priority and send data to other three queue based on priorities. Q1 is higher priority and data will insert is having priority 1 and other side Q2 will contain data priority is 2 and Q3 priority is 3.

If dequeue the data then must start from Q1 then Q2 then Q3. As the principal of priority queue is higher priority release it. So, Q1, Q2 and Q3 are set to their priorities, smaller number higher priority.

Enqueue based on priorities must be only 1 to 3.