Created AVL Tree
AVL tree is a self balancing tree which is searching time O(logn) and its satisfy the condition each node is height is either -1, 0 or 1 that is a balanced tree after than if any root node has height is 2 then it will heavy on left side of three or if node height has -2 more than it will heavy on right side of the tree. AVL Tree is a perfect binary search tree.